Phase 1: extract features/minimap.js and features/custom-css.js
Two self-contained features (no shared app state):
- minimap.js: updateMinimap / updateMinimapViewport — operate on the passed SVG
  element / panzoom instance / wrapper, using getSvgNaturalDimensions from
  core/utils and the fullscreen minimap DOM.
- custom-css.js: applyCustomCss — injects the desktop user stylesheet; its
  customStyleEl handle is now module-private.

Pure refactor. main.js ~2,324 -> ~2,245 lines. Build + lint clean; 297 tests pass.
